Tests/LibWeb/Text/expected/selectionchange-on-textarea.txt

selectionchange event dispatched on textarea should bubble: true
Selected range: 6 - 10
